# backend/docs/openapi/openapi.yaml
...
openapi: 3.1.0
paths:
...
  /transactions/by-account/{accountId}:
    get:
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      - name: status
        in: query
        schema:
          $ref: "#/components/schemas/TransactionStatus"
      - name: type
        in: query
        schema:
          $ref: "#/components/schemas/TransactionType"
      responses:
        "200":
          description: OK
      summary: Get Transactions Paged
      tags:
      - Transaction Controller
  /transactions/deposit:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Deposit
      tags:
      - Transaction Controller
  /transactions/deposit/agency:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Deposit By Agency
      tags:
      - Transaction Controller
  /transactions/payment:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Pay
      tags:
      - Transaction Controller
  /transactions/pix:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Pix Transfer
      tags:
      - Transaction Controller
  /transactions/receipt/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Get Receipt
      tags:
      - Transaction Controller
  /transactions/transfer:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Transfer
      tags:
      - Transaction Controller
  /transactions/transfer/agency:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Transfer By Agency
      tags:
      - Transaction Controller
  /transactions/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Get By Id
      tags:
      - Transaction Controller
...
components:
  schemas:
...
    TransactionRequestDTO:
      type: object
      properties:
        fromAccountId:
          type: integer
          format: int64
        toAccountId:
          type: integer
          format: int64
        agencyNumber:
          type: string
        accountNumber:
          type: string
        pixKey:
          type: string
        amount:
          type: string
        type:
          type: string
        description:
          type: string
    TransactionStatus:
      type: string
      enum:
      - COMPLETED
      - FAILED
    TransactionType:
      type: string
      enum:
      - PIX
      - DEPOSIT
      - TRANSFER
      - LOAN_DISBURSEMENT
      - BILL_PAYMENT
      - LOAN_INSTALLMENT_PAYMENT
      - PAYMENT
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8080/realms/crediflow/.well-known/openid-configuration
      description: Authentication
...
info:
  title: backend API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8082
  description: Auto generated value
- url: http://0.0.0.0:8082
  description: Auto generated value